Python WIP
The basic cd and ls functionality is now there, and the tab completion
functionality is in progress.

Readline + history functionality comes out of the box, which is nice,
though I haven't yet set up a persisted history file.

Tab completion is a tad buggy right now but is getting there.