Make ls command a little more sophisticated
Allow -l and -1 flags to be passed to ls, rather than the simpler
approach of just having ll as an alias for what would be ls -l in unix
utils.

Use a modified argument parser to handle those arguments robustly; this
also fixes an error where you'd get an unusual error message if you
provided too many arguments to the ls command, for example.