Fix confusing error on invalid bookmark key
We were "cleaning" the key by setting it to None if it didn't match the
pattern, and then complaining it was None, which isn't very helpful.

We'll never want to modify a new key being added, only explain why the
one picked was invalid, so change the functionality to be validate_key
instead and just check it against the regex.