Issue #25: paginate ls results
There was an existing TODO for this one, essentially it's limiting
results to 1000 while not paginated, so use a paginator to get all the
results.

There's an obvious caveat here about flat structures with very large
numbers of keys, so in future it may be better to limit this to a
configurable max key count and maybe even provide a way of paginating it
manually, but for now if you're trying to ls a prefix you expect to
contain millions of keys, expect a lot of s3 calls to see the results.