Explicitly use GNU readline
On some platforms like MacOS, and for some reason when running via uv,
the backend for the readline module is actually libedit instead of
readline.

Libedit has a completely different API so history files will be
corrupted and tab completion won't work with libedit currently.

Explicitly add gnureadline as a dependency and use it regardless of
python's default readline backend

In future I'll try to support both GNU readline and libedit backends,
but there's a fair number of edgecases to cover, including python
providing different methods of checking the backend in use between
python 3.12 and 3.13, so that's a job for another day.