Fix root-level completion
Effectively the problem here is that we're trying to turn a path, which
might be relative or absolute, partial or whole, into an S3Path
representing what actual path is being referred to, and then giving that
to ls.

The problem with that is that a partial path doesn't really represent an
S3Path. There's a bool flag used by the completion to represent that
it's not necessarily complete path if it didn't end with /, but then
it's getting confused when we're directly inside a bucket because it
thinks the bucket name might be incomplete, so it looks for buckets.

Resolve this by checking if we're currently inside a bucket in the
completer, and passing an extra flag to ignore buckets if we are --
because we're not looking for buckets.

Add __repr__ methods to some models to improve the way they're logged