Fix bucket search logic
The fix here wasn't quite right as it assumed the search path wasn't
absolute -- update this and instead determine whether buckets should be
searched by checking the path's bucket and path, and whether it ends
with a slash.

However, this puts us right back where we were before, with an edge case
bug while trying to complete an empty path while already in a bucket --
so resolve this by also considering the path a fragment if the "partial"
string is completely empty -- which can only represent a relative path
whic hhasn't been started at all yet, covering this edgecase.